tools/requirements.txt
airtable-python-wrapper
PyYAML
python-slugify